[#87] Testing backward compatible evaluate_bspline 3

+++ bsplines2d/_class02_bsplines_rect.py
@@ -40,7 +40,7 @@ else:
             t[None, :],
             np.array([t.size], dtype=np.int32),
             np.array([k], dtype=np.int32),
-            np.array([nu], dtype=int),
+            np.array([nu], dtype=np.int32),
             extrapolate, c, num_c_tr,
             strides_c1, indices_k1d, out,
         )
